task/344888: Сhecked the layout on the user friendly interface, moved the validation to the traits file.
